+++ CHANGELOG.rst
@@ -44,9 +44,12 @@ Major Features and Improvements
 
 - improved ``compute_errors`` in speed by caching values and the reliability
   by making the solution unique.
+- increased stability for large datasets with a constant subtraction in the NLL
 
 Breaking changes
 ------------------
+- NLL (and extended) subtracts now by default a constant value. This can be changed with a new `options` argument.
+  COMPARISON OF DIFFEREN NLLs (their absolute values) fails now! (flag can be deactivated)
 - BFGS (from TensorFlow Probability) has been removed as it is not working properly. There are many alternatives
   such as ScipyLBFGSV1 or NLoptLBFGSV1
 - Scipy (the minimizer) has been removed. Use specialized `Scipy*` minimizers instead.